Improve structure, fixes
