+++ factgenie/templates/crowdsourcing/annotate_header.html
@@ -19,4 +19,6 @@
   <script src="{{ host_prefix }}/static/js/lib/backbone.marionette.min.js"></script>
   <script src="{{ host_prefix }}/static/js/lib/underscore.string.min.js"></script>
   <script src="{{ host_prefix }}/static/js/lib/ypet.js"></script>
+
+  {% include 'crowdsourcing/custom.html' %}
 </head>
