+++ factgenie/templates/include/actions_modal.html
@@ -45,8 +45,8 @@
                                 <small class="text-muted">View the statistics of the annotation
                                     spans.</small>
                             </p>
-                            <a href="{{ host_prefix }}/analyze/detail/{{ campaign_id }}?source={{ metadata.mode }}"
-                                type="button" class="btn btn-primary"><i class="fa fa-area-chart"></i>
+                            <a href="{{ host_prefix }}/analyze/detail/{{ campaign_id }}" type="button"
+                                class="btn btn-primary"><i class="fa fa-area-chart"></i>
                                 View
                                 statistics</a>
                             <hr>
