+++ factgenie/templates/include/add_dataset_modal.html
@@ -14,8 +14,8 @@
                                 data-bs-toggle="pill">Basic loaders</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="#tab-download-dataset" data-bs-toggle="pill">Download a
-                                dataset</a>
+                            <a class="nav-link" href="#tab-download-dataset" data-bs-toggle="pill">Example
+                                datasets</a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link" href="#tab-custom-dataset" data-bs-toggle="pill">Custom dataset</a>
@@ -83,7 +83,8 @@
                     </div>
                     <div id="tab-download-dataset" role="tabpanel" class="tab-pane p-3" tabindex="1">
                         <div class="alert alert-info blue-link" role="alert">
-                            <p>You can download a dataset (optionally also along with model outputs and annotations)
+                            <p>You can download example datasets (optionally also along with model outputs and
+                                annotations)
                                 from the list below.</p>
 
                             <p>After downloading the dataset, it will appear in the list of local datasets.</p>
