+++ factgenie/templates/pages/analyze.html
@@ -55,8 +55,7 @@
               {% for campaign_id, campaign in campaigns.items() %}
               <tr>
                 <td><img src="{{ host_prefix }}/static/img/{{ campaign.metadata.mode }}.png" style="max-width: 18px;"
-                    class="heading-img-inline"> <a
-                    href="{{ host_prefix }}/analyze/detail/{{ campaign.metadata.id }}?source={{ campaign.metadata.mode }}"
+                    class="heading-img-inline"> <a href="{{ host_prefix }}/analyze/detail/{{ campaign.metadata.id }}"
                     class="blue-link">{{
                     campaign.metadata.id }}</a></td>
                 <td>{{ campaign.metadata.created }}</td>
