+++ factgenie/templates/pages/crowdsourcing_detail.html
@@ -67,7 +67,7 @@
           </a>
           <a type="button" class="btn btn-outline-secondary" data-bs-toggle="modal"
             data-bs-target="#config-modal-{{ campaign_id }}"><i class="fa fa-cog"></i> Show configuration</a>
-          <a href="{{ host_prefix }}/analyze/detail/{{ campaign_id }}?source=crowdsourcing" type="button"
+          <a href="{{ host_prefix }}/analyze/detail/{{ campaign_id }}" type="button"
             class="btn btn-outline-secondary"><i class="fa fa-area-chart"></i> View statistics</a>
           <a type="button" class="btn btn-outline-secondary" data-bs-toggle="modal"
             data-bs-target="#campaign-help-modal"><i class="fa fa-question-circle">
@@ -161,7 +161,7 @@
             <p>Use the <b>Download annotations</b> button to download the annotations
               collected so far.</p>
             <p>See the <a class="blue-link"
-                href="https://github.com/ufal/factgenie/wiki/04-Crowdsourcing-Annotations">crowsourcing
+                href="https://github.com/ufal/factgenie/wiki/Crowdsourcing-Annotations">crowsourcing
                 annotation tutorial</a> on the project wiiki for more information.</p>
           </div>
         </div>
