+++ factgenie/templates/pages/llm_campaign_new.html
@@ -118,7 +118,7 @@
                   <div class="mb-2">
                     <small class="form-text text-muted">Prompt for the model. See the instructions for configuring the
                       prompt at <a
-                        href="https://github.com/ufal/factgenie/wiki/03-LLM-Annotations#-configuring-the-prompt">factgenie
+                        href="https://github.com/ufal/factgenie/wiki/LLM-Annotations#-configuring-the-prompt">factgenie
                         wiki</a>.
                     </small>
                   </div>
